+++ src/AppInstallerCLICore/AppInstallerCLICore.vcxproj
@@ -201,7 +201,6 @@
     <ClInclude Include="ExecutionContext.h" />
     <ClInclude Include="ExecutionProgress.h" />
     <ClInclude Include="ExecutionReporter.h" />
-    <ClInclude Include="AppInstallerCLICoreExecute.h" />
     <ClInclude Include="Invocation.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="Public\AppInstallerCLICore.h" />
