+++ src/AppInstallerCLICore/Command.h
@@ -112,4 +112,6 @@ namespace AppInstaller::CLI
 
         return result;
     }
+
+    int Execute(Execution::Context& context, std::unique_ptr<Command>& command);
 }
